monitoring: make -multi-instance-groupings only generate multi-instance stuffs
